refactor: extract grid tree column logic to a separate mixin
