+++ packages/grid/src/vaadin-grid-tree-column-mixin.d.ts
@@ -1,0 +1,19 @@
+/**
+ * @license
+ * Copyright (c) 2016 - 2023 Vaadin Ltd.
+ * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
+ */
+
+import type { Constructor } from '@open-wc/dedupe-mixin';
+import type { GridColumn } from './vaadin-grid-column';
+
+export declare function GridTreeColumnMixin<TItem, T extends Constructor<GridColumn>>(
+  superclass: T,
+): Constructor<GridTreeColumnMixinClass<TItem>> & T;
+
+export declare class GridTreeColumnMixinClass<TItem> {
+  /**
+   * JS Path of the property in the item used as text content for the tree toggle.
+   */
+  path: string | null | undefined;
+}
